fix(interop): move setupOpenTelemetry to OpenTelemetryUtil to avoid Android compilation failure
